Move `defaultRadius` to `Radius.kt`.
