+++ designSystem/src/main/java/com/example/designsystem/radius/Radius.kt
@@ -1,2 +1,12 @@
 package com.example.designsystem.radius
 
+import androidx.compose.ui.unit.dp
+
+val defaultRadius = MovioRadius(
+    extraSmall = 4.dp,
+    small = 8.dp,
+    medium = 16.dp,
+    large = 20.dp,
+    xLarge = 24.dp,
+    xxLarge = 32.dp
+)
